phân quyền và fix linh tinh

+++ app/controllers/Auth/UserController.php
@@ -6,7 +6,7 @@ use App\controllers\Controller;
 use App\SessionGuard as Guard;
 use App\Models\TaiKhoan;
 
-class UserController extends Controller 
+class UserController extends Controller
 {
     public function __construct()
     {
@@ -19,8 +19,10 @@ class UserController extends Controller
 
     public function index()
     {
-        
-        $this->sendPage('Auth/user' ,[
+        if(!Guard::TaiKhoan()->isAdmin()){
+            redirect('/TongQuan', ['message' => 'Xin lỗi bạn cần có quyền admin để sử dụng chức năng này']);
+        }
+        $this->sendPage('Auth/user', [
             'user' => TaiKhoan::get()
         ]);
     }
@@ -30,28 +32,31 @@ class UserController extends Controller
         if (!$user)
             redirect('/User', ['message' => 'Tài khoản không tồn tại']);
 
-            $data = [
-                'user' => $user,
-                
-                'errors' => session_get_once('errors'),
-                'message' => session_get_once('message'),
-                'old' => $this->getSavedFormValues()
-            ];
+        $data = [
+            'user' => $user,
+
+            'errors' => session_get_once('errors'),
+            'message' => session_get_once('message'),
+            'old' => $this->getSavedFormValues()
+        ];
         $this->sendPage('/Auth/edit', $data);
     }
     protected function filterUserData(array $data)
     {
         return [
             'tennv' => $data['hoten'] ?? null,
-            
+
             'diachi' => $data['diachi'] ?? null,
             'gioitinhnv' => $data['gioitinh'] ?? null,
             'role' => $data['vaitro'] ?? null
-            
+
         ];
     }
     public function update($id)
     {
+        if(!Guard::TaiKhoan()->isAdmin()){
+            redirect('/TongQuan', ['message' => 'Xin lỗi bạn cần có quyền admin để sử dụng chức năng này']);
+        }
         // Tìm sản phẩm theo id
         $user = TaiKhoan::find($id);
 
@@ -62,7 +67,7 @@ class UserController extends Controller
             $this->saveFormValues($data);
             redirect("/User/edit/$id", ['errors' => $model_errors]);
         }
-        
+
 
         $fillData = [
             'tennv' => $data['tennv'],
@@ -78,61 +83,66 @@ class UserController extends Controller
     }
     public function destroy($id)
     {
+        if(!Guard::TaiKhoan()->isAdmin()){
+            redirect('/TongQuan', ['message' => 'Xin lỗi bạn cần có quyền admin để sử dụng chức năng này']);
+        }
         $user = TaiKhoan::find($id);
 
         if (!$user)
             redirect('/User', ['message' => 'Tài khoản không tồn tại']);
 
-        
 
-        
+
+
         $user->delete();
 
         redirect('/User', ['message' => 'Xóa tài khoản thành công']);
     }
 
-    public function changePass(){
-        $this->sendPage('Auth/changepass');
-    }
+    // public function changePass()
+    // {
+    //     $this->sendPage('Auth/changepass');
+    // }
     protected function filterUserpass(array $data)
     {
         return [
             'matkhauht' => $data['matkhauht'] ?? null,
             'matkhaumoi' => $data['matkhaumoi'] ?? null,
             'matkhaumoicheck' => $data['matkhaumoicheck'] ?? null
-            
-            
+
+
         ];
     }
     public function editpass()
     {
-        $user= Guard::TaiKhoan();
-        if(!$user){
+        $user = Guard::TaiKhoan();
+        if (!$user) {
             $this->sendNotFound();
         }
         $data = [
-            
+
             'errors' => session_get_once('errors'),
-            'user'=>$user,
-            'messages'=>session_get_once('messages')
+            'user' => $user,
+            'messages' => session_get_once('messages')
         ];
 
         $this->sendPage('/Auth/changepass', $data);
     }
 
-    public function updatePass(){
-        $user= Guard::TaiKhoan();
-        if(!$user){
+    public function updatePass()
+    {
+        $user = Guard::TaiKhoan();
+        if (!$user) {
             $this->sendNotFound();
         }
-        $data=$this->filterUserpass($_POST);
-        $model_errors=TaiKhoan::validatepass($data);
+        $data = $this->filterUserpass($_POST);
+        $model_errors = TaiKhoan::validatepass($data);
         // var_dump($model_errors);
         if (empty($model_errors)) {
             // Dữ liệu hợp lệ...
             $user->fill(['matkhau' => $data['matkhaumoi']]);
             $user->save();
-            
+
             redirect('/User/changepass', ['messages' => 'Thay đổi mật khẩu thành công.']);
         }
         redirect('/User/changepass', ['errors' => $model_errors]);
